requirements.txt
# pip install -r requirements.txt

# Base ----------------------------------------
matplotlib=3.6.2
numpy=1.23.5
opencv-python=4.9.0.80
Pillow=9.3.0
PyYAML=6.0
requests=2.28.2
scipy=1.1.10
scikit-learn=1.3.2
# torch>=1.9.0
# torchvision>=0.8.1
tqdm=4.65.2

# Logging -------------------------------------
tensorboard=2.12.2
# wandb


# Plotting ------------------------------------
pandas=1.5.2
seaborn=0.12.2

# Export --------------------------------------
# coremltools>=4.1  # CoreML export
# onnx>=1.9.0  # ONNX export
onnx-simplifier>=0.3.6  # ONNX simplifier
# scikit-learn==0.19.2  # CoreML quantization
# tensorflow>=2.4.1  # TFLite export
# tensorflowjs>=3.9.0  # TF.js export

# Extras --------------------------------------
# albumentations>=1.0.3
# Cython  # for pycocotools https://github.com/cocodataset/cocoapi/issues/172
# pycocotools>=2.0  # COCO mAP
# roboflow
thop  # FLOPs computation
